flasher: Created a Firmware structure for describing probe firmware and moved the type determination logic into a new function
